refactor(ios): split SystemOnnxGenAiRuntime.swift by concern
930-line file bundled 8 unrelated private types alongside the public
class, hard to navigate/review. Split into decoders, sampling, tensor
execution, session loading, KV-cache detection, and model resolution
files; no public API or behavior change. Adds SamplingConfig unit
tests, previously uncovered.
